refactor(scheduler): split scheduled task command layers
